Fix Sum Calls
Added a Fix to match total calls with identified calls. This fix is not
a final fix. It seems that Queue Calls are duplicated in the total
stats. Totally weird stuff, still checking that issue. Just for now I'll
take those calls to being processed even if they are duplicated.